==> 07_burroughs.Rmd <==
---
title: 'Burroughs Creek (Shoshone National Forest) WBP Planting Microsite Analysis'
subtitle: "Draft v2"
author: 
  - Stephen Huysman, shuysman@gmail.com
email: "shuysman@gmail.com"
date: "`r Sys.Date()`"
bibliography: library.bib
csl: global-ecology-and-biogeography.csl
link-citations: true
output:
  bookdown::html_document2:
    theme: cerulean
    number_sections: true
    toc: true
---

```{css, echo = FALSE}
/* body .main-container { */
/*     max-width: 2000px !important; */
/*     width: 1080px !important; */
/* } */
/* body { */
/*     max-width: 2000px !important; */
/* } */
```

``` {r setup, include = FALSE}
library(tidyverse)
library(sf)
library(terra)
library(tidyterra)
library(ggnewscale)
library(ggspatial)
library(ggpubr)
library(kableExtra)

knitr::opts_chunk$set(
  echo = F,
  warning = F,
  message = F,
#  fig.height = 8,
#  fig.width = 10,
  fig.path = paste0("/home/steve/OneDrive/core_areas/img/burroughs/")
)

options(digits = 2)
terraOptions(verbose = FALSE,
             memfrac = 0.9)

elev <- rast("./data/burroughs_creek_USGS1m_clipped_nad83.tif") 
aspect <- rast("./data/planting_unit_aspect.tif")
slope <- rast("./data/planting_unit_slope.tif")
shade <- rast("./data/burroughs_creek_USGS1m_clipped_hillshade_nad83.tif")
soil <- rast("./data/soil_whc_025_aligned_clipped.tif") %>% round(2) %>% clamp(lower = 0, values = NA)

pal_greys <- hcl.colors(1000, "Grays")

units <- st_read("./data/units/Possible_WBP_Units_Nancy_Updated.shp") %>% st_cast("MULTIPOLYGON")

##  Generate fs roads gpkg with only burroughs roads
## fs_roads <- st_read("./data/fs_roads/S_USA.RoadCore_FS.shp")

## elev2 <- project(elev, crs(fs_roads))
## bbox <- st_bbox(elev2) %>% st_as_sfc()
## fs_roads <- st_intersection(fs_roads, bbox)

## st_write(fs_roads, "./data/fs_roads_burroughs.gpkg")

fs_roads <- st_read("./data/fs_roads_burroughs.gpkg")

stats <- data.frame(elev_mean = zonal(elev, vect(units), fun = "mean")[[1]],
                         elev_max = zonal(elev, vect(units), fun = "max")[[1]],
                         elev_min = zonal(elev, vect(units), fun = "min")[[1]],
                         slope_mean = zonal(slope, vect(units), fun = "mean", na.rm = TRUE)[[1]],
                         slope_max = zonal(slope, vect(units), fun = "max", na.rm = TRUE)[[1]],
                         slope_min = zonal(slope, vect(units), fun = "min", na.rm = TRUE)[[1]],
                    aspect_mean = zonal(aspect, vect(units), fun = "mean", na.rm = TRUE)[[1]],
                    aspect_max = zonal(aspect, vect(units), fun = "max", na.rm = TRUE)[[1]],
                    aspect_min = zonal(aspect, vect(units), fun = "min", na.rm = TRUE)[[1]]
                    )

units <- bind_cols(units, stats)
```

``` {r plot-settings, include = FALSE}
## Change from baseline plots
aet_change_scale <-
    scale_fill_viridis_c(option = "G")#, limits = c(0, 300))

cwd_change_scale <-
    scale_fill_viridis_c(option = "F")#, limits = c(0, 300))

aet_scale <-
    scale_fill_viridis_c(option = "D", limits = c(72, 631))

cwd_scale <-
  scale_fill_viridis_c(option = "B", limits = c(0, 400))

aet_sd_scale <-
    scale_fill_viridis_c(option = "A", limits = c(7, 101))

cwd_sd_scale <-
    scale_fill_viridis_c(option = "C", limits = c(0, 97))
```

# Introduction
This document is a **work in progress (WIP)** and **all data are preliminary**!

This report presents historical (1979-2022) and projected (2006-2099)
water balance for the Burroughs Creek site (Figure
\@ref(fig:planting-units)) in the Shoshone National Forest.  This
location was selected to plant whitebark pine (*Pinus albicaulis*)
seedlings in summer 2024.  The site previously burned in August 30,
2013 in the Burroughs wildfire ([MTBS ID
`wy4368810966320130830`](wy4368810966320130830_map.pdf)).  The
vegetation composition of the site prior to the burn is unknown, but
is assumed to have sustained populations of whitebark pine.  Current
satellite imagery shows that the Burroughs wildfire cleared most trees
from the site, with some remnant trees persisting in sheltered
locations.

This report applies the climatic water balance to the Burroughs Creek
site at a fine spatial scale to make recommendations of ideal planting
locations. Potential Evapotranspiration (PET) represents the potential
for plant growth given available energy if not limited by water
availability.  Actual Evapotranspiration (AET) is a measure of total
evaporative loss to the atmosphere, and is well-correlated with
measures of plant growth such as primary productivity.  The difference
between PET and AET is the Climatic Water Deficit (CWD), which
measures evaporative demand unmet by available water.  AET represents
the length and magnitude of growing conditions favorable to plants
while CWD provides a measure of drought stress.  The bivariate
relationship between AET and CWD is a robust tool to characterize the
biophysical environment of plants and has been shown to be
well-correlated with vegetation distributions across ten orders of
spatial magnitude
[@stephensonClimaticControlVegetation1990;@stephensonActualEvapotranspirationDeficit1998].

Fine-scale variation in topography and soil features has been shown to
strongly affect modeled water balance
[@dyerGISBasedWaterBalance2019]. These fine-scale variations in water
balance are intuitively understood by planters, who often select
planting locations based on microsite features such as presence of
logs or boulders that provide shade, decreasing evaporative demands.
Our objective was to use historical and projected climate data to
model the climatic water balance of the Burroughs Creek site using the
highest resolution topographic and soil survey data available to
understand fine-scale spatial variation in the climatic water balance
across the site in order to facilitate identification of suitable
planting microsites and guide planting of whitebark pine seedlings.
As whitebark pine is a long-lived tree species that reaches
reproductive age around 50 years in ideal conditions, successful
whitebark plantings require a consideration of both present and future
climatic conditions.

In general, desirable planting locations will have higher AET, to
promote rapid plant establishment and growth, and lower CWD, to
minimize drought stress and fire risk
[@thomaWaterBalanceIndicator2020].  Depending on management objects,
optimal site characteristics can vary and these results can be
interpreted in different ways.  For example, if management goals are
to:

1. Maximize successful establishment of seedlings - planters should
       choose planting locations that maximize AET to achieve higher
       seedling growth rates and minimize CWD to reduce drought
       stress.  Previous work has identified fastest growth rates in
       whitebark seedlings above an annual AET around 350 mm
       [@laufenbergBiophysicalGradientsPerformance2020].  Although
       direct comparisons with this work cannot be made without bias
       correction between the different climate data sets used, in
       general higher AET can be assumed to correlate with faster
       seedling growth and establishment.
2. Grow trees to cone-bearing age as quickly as possible - choose
       planting locations that maximize AET over the duration needed
       for trees to reach reproductive age, approximately 25-50 years.
3. Maximize longevity of adult trees - select planting locations that
       minimize interannual variability in stressors.  It is thought
       that trees can survive unfavorable conditions if they are
       adapted to them, and extreme swings in climatic conditions
       cause mortality in trees that are not adapted to those extremes
       (M. Tercek, personal communication, June 4, 2024).

Color scales used for maps are kept the same for each variable (AET,
CWD) throughout this report, so direct comparisons can be made between
maps for the same variable made for different periods of time. 

Images can be click to open full-size in a new browser tab.

## Topography

```{r planting-units, fig.cap = "Burroughs Creek Planting Units, with hillshade and USFS roads.", fig.height = 8, fig.width = 10, out.width = '100%', warning = FALSE, message = FALSE, echo = FALSE}
ggplot() +
    geom_spatraster(data = shade, show.legend = FALSE) +
##    geom_spatraster(data = subset(tmmn$daily_minimum_temperature, 1)) +
    scale_fill_gradientn(colors = pal_greys, na.value = NA) +
    #geom_sf(data = fs_roads, color = "orange", lwd = 1.5, alpha = 0.5) +
    #geom_sf_text(data = fs_roads, aes(label = ID), alpha = 0.9) +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), alpha = 0.4) +
  #labs(title = "Burroughs WBP Planting Units Hillshade") +
  labs(x = "", y = "") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))
```

``` {r planting-unit-statistics,  warning = FALSE, message = FALSE, echo = FALSE}
knitr::kable(
    select(as_tibble(units),
        Unit,
        Description = DESCRIPTIO,
        Acres = ACRES,
        elev_mean,
        elev_max,
        elev_min,
        slope_mean,
        slope_max,
        slope_min,
        aspect_mean,
        aspect_max,
        aspect_min
    ),
    digits = 2,
    caption = "Topographic Statistics of Burroughs Creek planting units")
```

##  Soils
``` {r soils, fig.height = 8, fig.width = 10, out.width = '100%', warning = FALSE, message = FALSE, echo = FALSE, fig.cap = "Burroughs Creek soil water holding capacity (WHC, mm).  WHC is taken from the SSURGO available water storage at 25 cm soil depth layer"}
ggplot() +
  geom_spatraster(data = soil) +
  #labs(title = "Burroughs WBP Planting Unit Soil WHC (mm) to 25cm soil depth") +
  labs(x = "", y = "", fill = "WHC (mm)") +
  scale_fill_viridis_b() +
  geom_sf(data = units, fill = NA, lwd = 1.5) +
  geom_sf_label(data = units, aes(label = Unit), fill = NA) +
  annotation_north_arrow(aes(location = "tr")) +
  annotation_scale(aes(location = "br"))
```

Soils range from `r global(soil, min, na.rm = TRUE)[[1]]` to `r global(soil, max, na.rm = TRUE)[[1]]` mm whc at 25 cm soil depth
(Figure \@ref(fig:soils)). Soil water capacity at 25 cm was selected
because the application of this report is for planting seedlings and
mature whitebark pine trees typically have shallow root systems not
typically penetrating more than 30cm of soil
(https://www.fs.usda.gov/database/feis/plants/tree/pinalb/all.html).
Soils in the WB2 unit hold approximately 8 mm more water on average
(Table \@ref(tab:soils-zonals)).

``` {r soils-zonals, echo = FALSE}
zonal(soil, vect(units), fun = "mean", na.rm = TRUE) %>%
    cbind(Unit = units$Unit, .) %>%
  knitr::kable(digits = 2, col.names = c("Unit", "Mean Soil WHC (mm)"), caption = "Average soil WHC (mm)")

zonal(soil, vect(units), fun = "max", na.rm = TRUE) %>%
    cbind(Unit = units$Unit, .) %>%
  knitr::kable(digits = 2, col.names = c("Unit", "Max Soil WHC (mm)"), caption = "Max soil WHC (mm)")

zonal(soil, vect(units), fun = "min", na.rm = TRUE) %>%
    cbind(Unit = units$Unit, .) %>%
    knitr::kable(digits = 2, col.names = c("Unit", "Mean Min WHC (mm)"), caption = "Min soil WHC (mm)")
```

The SSURGO soil automated report for the site can be read here [20240514_11420911928_90_Soil_Report.pdf](./20240514_11420911928_90_Soil_Report.pdf "20240514_11420911928_90_Soil_Report.pdf").

# Methods
A Thornthwaite-style water balance model
[@thornthwaiteApproachRationalClassification1948] was run at a
daily-time step for the area within the bounding box of the planting
units. Topographic and elevation data was provided by the 1-meter
LiDAR digital elevation model (DEM) provided by the USGS 3DEP
[@u.s.geologicalsurveyWYFEMAEast2020].  Slope and aspect were
calculated from the 3DEP DEM using tools provided in QGIS
[@QGIS_software].  

Soil water holding capacity (WHC) was estimated using the 'available
water storage from 0-25cm' (`aws025wta`) data provided by the SSURGO
database
[@naturalresourcesconservationserviceSoilSurveyGeographic2015].  A 1 m
raster was generated from the WHC data in QGIS using the 1 m DEM
raster as a reference for sampling grid cells.  

Historical climate data (maximum temperature, minimum temperature,
precipitation) from 1979-2022 was provided by the gridMET dataset at a
daily timestep and 4 km resolution
[@abatzoglouDevelopmentGriddedSurface2013].  Projected climate data
for the same variables was provided by the MACA data set at the same
timestep and spatial resolution
[@abatzoglouComparisonStatisticalDownscaling2012].  The MACA dataset
was downscaled using gridMET, so time series between 1979 and 2099 are
commensurate with no breaks between past and future. The historical
and projected climate grids were reprojected and resampled to the same
CRS, extent, and resolution as the 1 m DEM raster.

Four CMIP5 GCMs for RCP4.5 and RCP8.5 scenarios were selected to
bracket plausible future conditions \@ref(tab:GCM-table). GCM/Scenario
combinations that represented warm wet, hot wet, warm dry, and hot dry
conditions in Yellowstone National Park (representing conditions in
the larger GYE) were selected based on projected changes in annual
temperature and precipitation by 2050 (Data from Mike Tercek,
citation?).

Table (\#tab:GCM-table): GCM and emission scenarios representing Warm
Wet, Hot Wet, Warm Dry, and Hot Dry conditions, bracketing the range
of plausible climate futures at the site.

| Future   | GCM           | Scenario |
|----------|---------------|----------|
| Warm Wet | MRI-CGCM3     | RCP8.5   |
| Hot Wet  | CanESM2       | RCP8.5   |
| Warm Dry | MRI-CGCM3     | RCP4.5   |
| Hot Dry  | HadGEM2-CC365 | RCP8.5   |

The Burroughs Creek site is much smaller in scale than the 4 km grid
cells, and falls at the center of the intersection of four grid cells.
To remove spatial discontinuities in the modeled water balance for the
site, the grid cell with an elevation closest to the mean elevation
for the site was selected and climate data from this grid cell was applied across the site.  

The daily water balance at a 1m resolution was modeled using a
modified version of the script used to generate the NPS 1 km Gridded
Water Balance product for the coterminous United States
[@tercekHistoricalChangesPlant2021].  PET was calculated following @oudinEstimatingPotentialEvapotranspiration2010.
Annual sums of AET (mm) and Deficit (mm) were calculated for the
historical dataset as well as for each GCM/Scenario combination.
Ensemble averages for annual sums of AET and Deficit were created.

For comparisons with projected conditions, annual AET and CWD sums
between 2002-2022 were averaged to use as current baseline conditions.
Projections were summarized for the end-century (2076-2099) period.

# Results
## Climate space of WBP

Historic (2000-2019) AET and CWD annual means were taken from the NPS
1km gridded water-balance product [@tercekHistoricalChangesPlant2021]
for known WBP occurrences across WBP's range in CONUS using a white
pine blister rust monitoring dataset assembled from data provided by
collaborators in the NPS, USFS, and BLM (Manuscript in preparation).
Historic as well as end-of-century (2070-2099) ensemble projections of
AET and CWD under RCP4.5 and RCP8.5 emissions scenarios were also
taken for a point located in the Burroughs Creek planting site.
Figure \@ref(fig:climate-space) shows the relationship between AET and
CWD for these points where WBP is known to occur as well where the
Burroughs Creek site fits into this broad scale climate space.  In
this figure, WBP occurrences in the GYE are highlighted.

This analysis shows that historically, the Burroughs Creek site was on
average in the core of the AET/CWD climate-space that WBP inhabits in
the GYE, and likely provided favorable climate for the species.
However, ensemble averages for projections under RCP4.5 and RCP8.5
scenarios both show the site, on average, transitioning to a climate
that resembles sites ranging from at the edge of whitebark pine's
current habitat (RCP4.5) to experiencing combinations of AET and CWD
that are extreme compared to historical climate at any of the known
WBP occurrences included (RCP8.5).  Under either scenario, this site
is likely to transition to a climate that is extreme to uninhabitable
for WBP's climate preferences.  However, it is possible that
fine-scale terrain variation can create local water balances that
ameliorate these projected climate extremes creating favorable water
balances for WBP.  The 1m Water Balance analysis presented here
attempts to identify these fine-scale terrain features that could both
create favorable local climates for planting now as well as buffer planted
trees from extreme climates in the future when they mature. Given the
unfavorable projections of this site under projected climate
scenarios, it is imperative that plantings performed now take
advantage of the effects of fine-scale terrain features on local
climates.

It should be noted that the analysis presented in figure \@ref(fig:climate-space) was generated with a 1 km gridded water balance model and magnitudes from this analysis should not be compared with magnitudes generated in the finer-scale 1 m water balance.  Relative patterns within models at a given scale are comparable, for example both models can tell if a position on the landscape is drier than another.  However, the annual average AET and CWD values between the models at different scales are not comparable.  Additionally, the 1 km gridded water balance model used soil WHC at 100 cm, while soil WHC at 25 cm was used here in order to simulate the growing environment of seedlings or shallow-rooted mature trees, which further compounds the issue of comparing absolute values between the models.  

``` {r climate-space, fig.height = 5, fig.width = 5, warning = FALSE, message = FALSE, echo = FALSE, fig.cap = "Bioclimatic niche space of WBP.  WBP occurrence points taken from NPS white pine blister rust monitoring data set.  WBP occurrences within the Greater Yellowstone Ecosystem are highlighted in light blue. AET and CWD values are sampled from the 1 km. NPS Gridded Water Balance model summary layers (http://screenedcleanedsummaries.s3-website-us-west-2.amazonaws.com/).  Historical and RCP4.5/RCP8.5 projections for AET and CWD for a point in the Burroughs Creek site (43.7061975, -109.6778113) were also sampled from the NPS Gridded Water Balance model summary layers and displayed using a + symbol. Projections are ensemble averages of CMIP5 models. Note that AET and CWD values in the NPS Gridded Water Balance product were calculated using different model parameters (soil WHC, topography) and should not be compared to the 1m water balance model presented here."}
#library(plotly)
library(gghighlight)

burroughs_ck <- c(-109.6778113, 43.7061975)

aet_summary_dir <- file.path("/media/smithers/shuysman/data/nps_gridded_wb/summary_layers/AET/")
cwd_summary_dir <- file.path("/media/smithers/shuysman/data/nps_gridded_wb/summary_layers/Deficit/")

aet <- rast(file.path(aet_summary_dir, "historical/V_1_5_annual_gridmet_historical_AET_2000_2019_annual_means_cropped_units_mm.tif"))
names(aet) <- "AET"
cwd <- rast(file.path(cwd_summary_dir, "historical/V_1_5_annual_gridmet_historical_Deficit_2000_2019_annual_means_cropped_units_mm.tif"))
names(cwd) <- "CWD"
aet_45 <- rast(file.path(aet_summary_dir, "rcp45/ensembles/ensemble_2070_2099_annual_rcp45_AET_units_mm.tif"))
names(aet_45) <- "AET_45"
cwd_45 <- rast(file.path(cwd_summary_dir, "rcp45/ensembles/ensemble_2070_2099_annual_rcp45_Deficit_units_mm.tif"))
names(cwd_45) <- "CWD_45"
aet_85 <- rast(file.path(aet_summary_dir, "rcp85/ensembles/ensemble_2070_2099_annual_rcp85_AET_units_mm.tif"))
names(aet_85) <- "AET_85"
cwd_85 <- rast(file.path(cwd_summary_dir, "rcp85/ensembles/ensemble_2070_2099_annual_rcp85_Deficit_units_mm.tif"))
names(cwd_85) <- "CWD_85"

wbp_points <- read_csv("/home/steve/OneDrive/whitebark/blister-rust/data/SITE_LOCATIONS.csv") %>%
  bind_rows(tibble(network = "burroughs", park = "burroughs", long = burroughs_ck[1], lat = burroughs_ck[2])) %>%
  drop_na(c("lat", "long")) %>%
  st_as_sf(coords = c("long", "lat"), crs = st_crs("EPSG:4326")) %>%
  st_transform(crs = st_crs(aet))

wbp_points <- extract(aet, wbp_points, bind = TRUE)
wbp_points <- extract(cwd, wbp_points, bind = TRUE)
wbp_points <- extract(aet_45, wbp_points, bind = TRUE)
wbp_points <- extract(cwd_45, wbp_points, bind = TRUE)
wbp_points <- extract(aet_85, wbp_points, bind = TRUE)
wbp_points <- extract(cwd_85, wbp_points, bind = TRUE)

## wbp_plot <- ggplot(wbp_points) +
##     geom_point(aes(x = CWD, y = AET, color = park)) +
##     labs(title = "WBP Bioclimatic Niche")
## ggplotly(wbp_plot)

ggplot(wbp_points) +
    geom_point(aes(x = CWD, y = AET, color = park)) +
    ## geom_point(aes(x = 179.89, 192.98), color = "blue", shape = 3, size = 10) +
    ## geom_text(aes(x = 179.89, 192.98), color = "blue", label = "historic") +
    ## geom_point(aes(x = 235.81, 247.61), color = "green", shape = 2, size = 10) +
    ## geom_text(aes(x = 235.81, 247.61), color = "green", label = "rcp4.5") +
    ## geom_point(aes(x = 306.32, 287.88), color = "red", shape = 2, size = 10) +
  ## geom_text(aes(x = 306.32, 287.88), color = "red", label = "rcp8.5") +
  geom_point(shape = "+", size = 12, data = filter(wbp_points, park == "burroughs"), aes(x = CWD, y = AET), color = "black") +
  geom_point(shape = "+", size = 12, data = filter(wbp_points, park == "burroughs"), aes(x = CWD_45, y = AET_45), color = "orange") +
  geom_point(shape = "+", size = 12, data = filter(wbp_points, park == "burroughs"), aes(x = CWD_85, y = AET_85), color = "red") +
  geom_text(data = filter(wbp_points, park == "burroughs"), aes(x = CWD, y = AET), label = "Historical", nudge_y = -10, size = 5) +
  geom_text(data = filter(wbp_points, park == "burroughs"), aes(x = CWD_45, y = AET_45), label = "4.5", nudge_y = -10, size = 5) +
  geom_text(data = filter(wbp_points, park == "burroughs"), aes(x = CWD_85, y = AET_85), label = "8.5", nudge_y = -10, size = 5) +
  labs(x = "Annual CWD (mm)", y = "Annual AET (mm)") +
  gghighlight(park %in% c("GYE", "burroughs"))
  #labs(title = "WBP Bioclimatic Niche (GYE)")


```

## Historical Water Balance

Overall, both AET and CWD increased across the site from 1982-2002 to 2002-2022 (Table \@ref(tab:climate-summary,), Figures \@ref(fig:historical-aet-plot), \@ref(fig:historical-cwd-plot)). 

``` {r historical, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 6, fig.width = 8}
gridmet_data_dir <- file.path("/media/smithers/shuysman/data/out/nps-wb/burroughs/sums/")

aet <- rast(file.path(gridmet_data_dir, "historical_gridmet_AET_annual_sum.nc")) |>
    clamp(lower = 0, values = NA) / 10
cwd <- rast(file.path(gridmet_data_dir, "historical_gridmet_Deficit_annual_sum.nc")) |>
    clamp(lower = 0, values = NA) / 10

arrow_w <- unit(.75, "cm")
arrow_h <- unit(.75, "cm")

### 1988 Drought Year
aet_1988 <- aet %>% subset(year(time(.)) == 1988)
cwd_1988 <- cwd %>% subset(year(time(.)) == 1988)

aet_2013 <- aet %>% subset(year(time(.)) == 2013)
cwd_2013 <- cwd %>% subset(year(time(.)) == 2013)

### Previous 20 years before current
aet_2002 <- aet %>% subset(year(time(.)) >= 1982 & year(time(.)) <= 2002)  %>% mean()
cwd_2002 <- cwd %>% subset(year(time(.)) >= 1982 & year(time(.)) <= 2002) %>% mean()

### Last 20 years
aet_2022 <- aet %>% subset(year(time(.)) >= 2002 & year(time(.)) <= 2022) %>% mean()
cwd_2022 <- cwd %>% subset(year(time(.)) >= 2002 & year(time(.)) <= 2022) %>% mean()

writeRaster(aet_2022, filename = "raster/baseline_aet_2002-2022.tif", overwrite = TRUE)
writeRaster(cwd_2022, filename = "raster/baseline_cwd_2002-2022.tif", overwrite = TRUE)

years = c(1988, 2002, 2022)
vars = c("aet", "cwd")
types = c("mean", "min", "max")

unit_stats <- tibble()
for (year in years) {
    for (var in vars) {
        for (type in types) {
            unit_stats <- bind_rows(unit_stats,
                               bind_cols(year = year,
                                         var = var,
                                         type = type,
                                         value = zonal(get(glue::glue("{var}_{year}")), vect(units), fun = type, na.rm = TRUE)[[1]]))
        }
    }
}

unit_stats <- bind_cols(unit = rep(units$Unit, nrow(unit_stats) / 2), unit_stats)

## unit_stats %>% pivot_wider(id_cols = c(unit, year, var),
##                             names_from = type,
##                            values_from = value) %>%
##     ggplot(aes(x = year, y = mean, color = unit)) +
##     geom_line() +
##     geom_ribbon(aes(ymin = min, ymax = max, fill = unit), linetype = "dashed", alpha = 0.1) +
##     ##geom_linerange(aes(x = year, y = mean, ymin = min, ymax = max), color = unit) +
##     facet_wrap(~var)

## unit_stats %>% pivot_longer(cols = starts_with(c("aet"))) %>%
##     ggplot() +
##     geom_line(aes(x = name, y = value, group = Unit, color = Unit)) +
##     labs(title = "Change in AET normals vs 1988 drought year")

## unit_stats %>% pivot_longer(cols = starts_with(c("cwd"))) %>%
##     ggplot() +
##     geom_line(aes(x = name, y = value, group = Unit, color = Unit)) +
##     labs(title = "Change in CWD normals vs 1988 drought year")

aet_2002_plot <- ggplot() +
    geom_spatraster(data = aet_2002, alpha = 1) +
    aet_scale + 
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "1982-2002", fill = "AET (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

cwd_2002_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data =cwd_2002, alpha = 1) +
    cwd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "1982-2002", fill = "CWD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))


aet_2022_plot <- ggplot() +
    geom_spatraster(data = aet_2022, alpha = 1) +
    aet_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2002-2022", x = "", y = "", fill = "AET (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))
##ggsave("burroughs_2002-2022_aet.png", width=16, height=12)


cwd_2022_plot <- ggplot() +
    geom_spatraster(data = cwd_2022, alpha = 1) +
    cwd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2002-2022", x = "", y = "", fill = "CWD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))
##ggsave("burroughs_2002-2022_deficit.png", width=16, height=12)



aet_1988_plot <- ggplot() +
    geom_spatraster(data = aet_1988, alpha = 1) +
    theme(legend.position="bottom") +
    aet_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "1988", x = "", y = "", fill = "AET (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))


cwd_1988_plot <- ggplot() +
    geom_spatraster(data = cwd_1988, alpha = 1) +
    theme(legend.position="bottom") +
    cwd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "1988", x = "", y = "", fill = "CWD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))
```

``` {r historical-aet-plot, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 6, fig.width = 14, out.width = '100%', fig.cap = "Historical AET (mm)"}
ggarrange(aet_2002_plot, aet_2022_plot, common.legend = TRUE) %>%
    annotate_figure(top = text_grob("Historical annual AET (mm)", size = 14, face = "bold"))
```

``` {r historical-cwd-plot, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 6, fig.width = 14, out.width = '100%', fig.cap = "Historical CWD (mm)"} 
ggarrange(cwd_2002_plot, cwd_2022_plot, common.legend = TRUE) %>%
    annotate_figure(top = text_grob("Historical annual CWD (mm)", size = 14, face = "bold"))
```



``` {r drought-year-plot, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 6, fig.width = 14, out.width = '100%', fig.cap = "Drought Year Conditions (1988)"}
ggarrange(aet_1988_plot, cwd_1988_plot) %>%
    annotate_figure(top = text_grob("Dought Year (1988)", size = 14, face = "bold"))
```

## Projected Water Balance

End-of-century (2075-2099) projections of AET and CWD are shown for
four GCMs from the MACA dataset bracketing the range of plausible
future climates at the site (table \@ref(tab:GCM-table)).  Projected AET shows increases across the Burroughs Creek landscape on average, with the Hot-Wet scenario showing the greatest increases (Figure \@ref(fig:projections-aet), Table \@ref(tab:climate-summary).  


``` {r projections-setup, warning = FALSE, message = FALSE, echo = FALSE}
current <- 2006:2024
near <- 2025:2049
mid <- 2050:2074
end <- 2075:2099

proj_dir <- file.path("/media/smithers/shuysman/data/out/nps-wb/burroughs/sums/")

load_gcm_mean <- function(filename) {
    r <- rast(filename) %>%
        subset(year(time(.)) %in% end) %>% ## Limit to end-century conditions
        mean() %>%
        clamp(lower = 0, values = FALSE) / 10 ## remove NAs and divide by ten (nps gridded wb is provided in units x 10)
    return(r)
}

load_gcm_sd <- function(filename) {
    r <- rast(filename) %>%
        stdev(na.rm = TRUE) / 10
    return(r)
}

aet_warm_wet <- load_gcm_mean(file.path(proj_dir, "MRI-CGCM3_rcp85_AET_annual_sum.nc"))
cwd_warm_wet <- load_gcm_mean(file.path(proj_dir, "MRI-CGCM3_rcp85_Deficit_annual_sum.nc"))

aet_hot_wet <- load_gcm_mean(file.path(proj_dir, "CanESM2_rcp85_AET_annual_sum.nc")) 
cwd_hot_wet <- load_gcm_mean(file.path(proj_dir, "CanESM2_rcp85_Deficit_annual_sum.nc"))

aet_warm_dry <- load_gcm_mean(file.path(proj_dir, "MRI-CGCM3_rcp45_AET_annual_sum.nc"))
cwd_warm_dry <- load_gcm_mean(file.path(proj_dir, "MRI-CGCM3_rcp45_Deficit_annual_sum.nc"))

aet_hot_dry <- load_gcm_mean(file.path(proj_dir, "HadGEM2-CC365_rcp85_AET_annual_sum.nc"))
cwd_hot_dry <- load_gcm_mean(file.path(proj_dir, "HadGEM2-CC365_rcp85_Deficit_annual_sum.nc"))

aet_warm_wet_sd <- load_gcm_sd(file.path(proj_dir, "MRI-CGCM3_rcp85_AET_annual_sum.nc"))
cwd_warm_wet_sd <- load_gcm_sd(file.path(proj_dir, "MRI-CGCM3_rcp85_Deficit_annual_sum.nc"))

aet_hot_wet_sd <- load_gcm_sd(file.path(proj_dir, "CanESM2_rcp85_AET_annual_sum.nc")) 
cwd_hot_wet_sd <- load_gcm_sd(file.path(proj_dir, "CanESM2_rcp85_Deficit_annual_sum.nc"))

aet_warm_dry_sd <- load_gcm_sd(file.path(proj_dir, "MRI-CGCM3_rcp45_AET_annual_sum.nc"))
cwd_warm_dry_sd <- load_gcm_sd(file.path(proj_dir, "MRI-CGCM3_rcp45_Deficit_annual_sum.nc"))

aet_hot_dry_sd <- load_gcm_sd(file.path(proj_dir, "HadGEM2-CC365_rcp85_AET_annual_sum.nc"))
cwd_hot_dry_sd <- load_gcm_sd(file.path(proj_dir, "HadGEM2-CC365_rcp85_Deficit_annual_sum.nc"))


## Worst Case Scenarios
writeRaster(aet_hot_wet, "raster/aet_hot_wet.tif", overwrite = TRUE)
writeRaster(cwd_hot_dry, "raster/cwd_hot_dry.tif", overwrite = TRUE)

```


``` {r, projections-aet, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 16, fig.width = 16, out.width = '100%', fig.cap = "Projected AET under selected GCMs representing hot/wet, hot/dry, warm/wet, and warm/dry conditions"}
aet_warm_wet_plot <- ggplot() +
  geom_spatraster(data = clamp(aet_warm_wet, lower = 1, values = NA), alpha = 1) +
  aet_scale +
  geom_sf(data = units, fill = NA, lwd = 1.5) +
  geom_sf_label(data = units, aes(label = Unit), fill = NA) +
  labs(title = "2075–2099 Warm-Wet (MRI-CGCM3 RCP8.5)", x = "", y = "", fill = "AET (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
  annotation_scale(aes(location = "br"))

cwd_warm_wet_plot <- ggplot() +
    geom_spatraster(data = clamp(cwd_warm_wet, lower = 1, values = NA), alpha = 1) +
    cwd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2075–2099 Warm-Wet (MRI-CGCM3 RCP8.5)", x = "", y = "", fill = "CWD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

aet_hot_wet_plot <- ggplot() +
    geom_spatraster(data = clamp(aet_hot_wet, lower = 1, values = NA), alpha = 1) +
    aet_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2075–2099 Hot-Wet (CanESM2 RCP8.5)", x = "", y = "", fill = "AET (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

cwd_hot_wet_plot <- ggplot() +
    geom_spatraster(data = clamp(cwd_hot_wet, lower = 1, values = NA), alpha = 1) +
    cwd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2075–2099 Hot-Wet (CanESM2 RCP8.5)", x = "", y = "", fill = "CWD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

aet_warm_dry_plot <- ggplot() +
    geom_spatraster(data = clamp(aet_warm_dry, lower = 1, values = NA), alpha = 1) +
    aet_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2075–2099 Warm-Dry (MRI-CGCM3 RCP4.5)", x = "", y = "", fill = "AET (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

cwd_warm_dry_plot <- ggplot() +
    geom_spatraster(data = clamp(cwd_warm_dry, lower = 1, values = NA), alpha = 1) +
    cwd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2075–2099 Warm-Dry (MRI-CGCM3 RCP4.5)", x = "", y = "", fill = "CWD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

aet_hot_dry_plot <- ggplot() +
    geom_spatraster(data = clamp(aet_hot_dry, lower = 1, values = NA), alpha = 1) +
    aet_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2075–2099 Hot-Dry (HadGEM2-CC365 RCP8.5)", x = "", y = "", fill = "AET (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

cwd_hot_dry_plot <- ggplot() +
    geom_spatraster(data = clamp(cwd_hot_dry, lower = 1, values = NA), alpha = 1) +
    cwd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "2075–2099 Hot-Dry (HadGEM2-CC365 RCP8.5)", x = "", y = "", fill = "CWD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))


ggarrange(
    aet_warm_wet_plot, aet_hot_wet_plot,
    aet_warm_dry_plot, aet_hot_dry_plot,
    common.legend = TRUE
)

```

CWD is also projected to increase across the landscape in all future scenarios examined, with the Hot-Dry scenario showing the greatest overall increase (Figure \@ref(fig:projections-cwd), Table \@ref(tab:climate-summary)).  CWD values in projections are similar to the extreme deficit experienced during the 1988 droughts, with the Hot-Dry scenario surpassing the 1988 CWD by 139 mm (409 vs 270 mm, Table \@ref(tab:climate-summary)). While this is a "worst-case" scenario, even in the best case (Warm-Dry), the average annual deficit is only 61 mm lower than the 1988 droughts (208 mm vs 270mm, Table \@ref(tab:climate-summary)).

``` {r, projections-cwd, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 16, fig.width = 16, out.width = '100%', fig.cap = "Projected CWD under selected GCMs representing hot/wet, hot/dry, warm/wet, and warm/dry conditions"}
ggarrange(
    cwd_warm_wet_plot, cwd_hot_wet_plot,
    cwd_warm_dry_plot, cwd_hot_dry_plot,
    common.legend = TRUE
)
```

``` {r, combined-aet-plot, fig.height = 9, fig.width = 11}
ggarrange(aet_1988_plot, aet_2022_plot,
          aet_hot_wet_plot, aet_warm_dry_plot,
          common.legend = TRUE
          )

```

``` {r, combined-cwd-plot, fig.height = 9, fig.width = 11}
ggarrange(cwd_1988_plot, cwd_2022_plot,
          cwd_warm_dry_plot, cwd_hot_dry_plot,
          common.legend = TRUE
          )
```

### Change in AET/CWD

The change in AET and CWD from baseline historical (2002-2022) conditions is calculated for AET and CWD projections.  This gives a relative measure of how extreme the climate could become compared to current conditions.  The Hot-Dry scenario shows the greatest increases in AET across the landscale (Figure \@ref(fig:projections-change-aet)).

``` {r, projections-change-aet, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 16, fig.width = 16, out.width = '100%', fig.cap = "Projected change in AET from historical conditions under selected GCMs representing hot/wet, hot/dry, warm/wet, and warm/dry conditions"}
aet_warm_wet_change <- aet_warm_wet - aet_2022
cwd_warm_wet_change <- cwd_warm_wet - cwd_2022

aet_hot_wet_change <- aet_hot_wet - aet_2022
cwd_hot_wet_change <- cwd_hot_wet - cwd_2022

aet_warm_dry_change <-  aet_warm_dry - aet_2022
cwd_warm_dry_change <- cwd_warm_dry - cwd_2022

aet_hot_dry_change <- aet_hot_dry - aet_2022
cwd_hot_dry_change <- cwd_hot_dry - cwd_2022

writeRaster(aet_hot_wet_change, "raster/aet_hot_wet_change.tif", overwrite = TRUE)
writeRaster(cwd_hot_dry_change, "raster/cwd_hot_dry_change.tif", overwrite = TRUE)

aet_warm_wet_change_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = aet_warm_wet_change, alpha = 1) +
    aet_change_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Warm Wet") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

aet_hot_wet_change_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = aet_hot_wet_change, alpha = 1) +
    aet_change_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Hot Wet") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

aet_warm_dry_change_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = aet_warm_dry_change, alpha = 1) +
    aet_change_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Warm Dry") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

aet_hot_dry_change_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = aet_hot_dry_change, alpha = 1) +
    aet_change_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Hot Dry") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

cwd_warm_wet_change_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = cwd_warm_wet_change, alpha = 1) +
    cwd_change_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Warm Wet") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

cwd_hot_wet_change_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = cwd_hot_wet_change, alpha = 1) +
    cwd_change_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Hot Wet") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

cwd_warm_dry_change_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = cwd_warm_dry_change, alpha = 1) +
    cwd_change_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Warm Dry") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

cwd_hot_dry_change_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = cwd_hot_dry_change, alpha = 1) +
    cwd_change_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Hot Dry") +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

ggarrange(
    aet_warm_wet_change_plot, aet_hot_wet_change_plot,
    aet_warm_dry_change_plot, aet_hot_dry_change_plot,
    common.legend = TRUE
)

```
 The Hot-Dry scenario also shows the greatest increases in CWD across the landscape, with Warm-Wet also showing similar increases (Figure \@ref(fig:projections-change-cwd)).
 
``` {r, projections-change-cwd, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 16, fig.width = 16, out.width = '100%', fig.cap = "Projected change in CWD from historical conditions under selected GCMs representing hot/wet, hot/dry, warm/wet, and warm/dry conditions"}
ggarrange(
    cwd_warm_wet_change_plot, cwd_hot_wet_change_plot,
    cwd_warm_dry_change_plot, cwd_hot_dry_change_plot,
    common.legend = TRUE
)
```


### Interannual Variability 

The standard deviation in annual AET and CWD are calculated as a measure of interannual variability (IAV).  This tells how much a pixel changes year to year.  Locations that do not change much on average can be desirable for tree longevity, as rapid shifts in climate can result in tree mortality (M. Tercek, personal communication, June 4, 2024).  The Hot-Dry and Warm-Wet scenarios show the highest IAV in AET (Figure: \@ref(fig:projections-sd-aet)).

``` {r, projections-sd-aet, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 9, fig.width = 11, fig.cap = "AET Interannual variability (standard deviation) in climatic conditions under projections from selected GCMs representing hot/wet, hot/dry, warm/wet, and warm/dry conditions"}
aet_warm_wet_sd_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = aet_warm_wet_sd, alpha = 1) +
    aet_sd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Warm Wet", x = "", y = "", fill = "SD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

aet_hot_wet_sd_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = aet_hot_wet_sd, alpha = 1) +
    aet_sd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Hot Wet", x = "", y = "", fill = "SD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

aet_warm_dry_sd_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = aet_warm_dry_sd, alpha = 1) +
    aet_sd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Warm Dry", x = "", y = "", fill = "SD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

aet_hot_dry_sd_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = aet_hot_dry_sd, alpha = 1) +
    aet_sd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Hot Dry", x = "", y = "", fill = "SD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

cwd_warm_wet_sd_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = cwd_warm_wet_sd, alpha = 1) +
    cwd_sd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Warm Wet", x = "", y = "", fill = "SD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

cwd_hot_wet_sd_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = cwd_hot_wet_sd, alpha = 1) +
    cwd_sd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Hot Wet", x = "", y = "", fill = "SD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

cwd_warm_dry_sd_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = cwd_warm_dry_sd, alpha = 1) +
    cwd_sd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Warm Dry", x = "", y = "", fill = "SD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

cwd_hot_dry_sd_plot <- ggplot() +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = cwd_hot_dry_sd, alpha = 1) +
    cwd_sd_scale +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), fill = NA) +
    labs(title = "Hot Dry", x = "", y = "", fill = "SD (mm)") +
    annotation_north_arrow(aes(location = "tr"), height = arrow_h, width = arrow_w) +
    annotation_scale(aes(location = "br"))

ggarrange(
    aet_warm_wet_sd_plot, aet_hot_wet_sd_plot,
    aet_warm_dry_sd_plot, aet_hot_dry_sd_plot,
    common.legend = TRUE
)
```

The Hot-Dry scenario shows the highest IAV in CWD (Figure: \@ref(fig:projections-sd-cwd)).


``` {r, projections-sd-cwd, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 9, fig.width = 11, fig.cap = "CWD Interannual variability (standard deviation) in climatic conditions under projections from selected GCMs representing hot/wet, hot/dry, warm/wet, and warm/dry conditions"}

ggarrange(
    cwd_warm_wet_sd_plot, cwd_hot_wet_sd_plot,
    cwd_warm_dry_sd_plot, cwd_hot_dry_sd_plot,
    common.legend = TRUE
)

```

## Planting Site Selection
An algorithm to select planting locations based on this data is
presented here.  The 1,000,000 pixels with the highest AET and
1,000,000 pixels with the lowest CWD are selected.  Pixels that
overlap between these layers are then selected, giving pixels that
maximize AET and minimize CWD. The final number of selected pixels is
around 200,000
or approximately 10-15% of the landscape.  

``` {r, best-plantings-setup1, eval = FALSE, include = FALSE, warning = FALSE, echo = FALSE, message = FALSE}

n_sites <- 1000000

### Historical
## (n_sites / ncell(elev)) * 100

highest_aet_historical <- selectHighest(aet_2022, n = n_sites)
lowest_cwd_historical <- selectHighest(cwd_2022, n = n_sites, low = TRUE)

best_sites_historical <- highest_aet_historical & lowest_cwd_historical

n_best_sites_historical <- zonal(best_sites_historical, best_sites_historical, length)[[2]]

## (n_best_sites / ncell(elev)) * 100

mean_aet_historical <- global(mask(aet_2022, highest_aet_historical), fun = "mean", na.rm = TRUE)
mean_cwd_historical <- global(mask(cwd_2022, lowest_cwd_historical), fun = "mean", na.rm = TRUE)
```

```{r, best-plantings-setup2, eval = FALSE, include = FALSE}

### Warm Wet

highest_aet_warm_wet <- selectHighest(aet_warm_wet, n = n_sites)
lowest_cwd_warm_wet <- selectHighest(cwd_warm_wet, n = n_sites, low = TRUE)

best_sites_warm_wet <- highest_aet_warm_wet & lowest_cwd_warm_wet

n_best_sites_warm_wet <- zonal(best_sites_warm_wet, best_sites_warm_wet, length)[[2]]

## (n_best_sites / ncell(elev)) * 100

mean_aet_warm_wet <- global(mask(aet_warm_wet, highest_aet_warm_wet), fun = "mean", na.rm = TRUE)
mean_cwd_warm_wet <- global(mask(cwd_warm_wet, lowest_cwd_warm_wet), fun = "mean", na.rm = TRUE)

### Hot Wet

highest_aet_hot_wet <- selectHighest(aet_hot_wet, n = n_sites)
lowest_cwd_hot_wet <- selectHighest(cwd_hot_wet, n = n_sites, low = TRUE)

best_sites_hot_wet <- highest_aet_hot_wet & lowest_cwd_hot_wet

n_best_sites_hot_wet <- zonal(best_sites_hot_wet, best_sites_hot_wet, length)[[2]]

## (n_best_sites / ncell(elev)) * 100

mean_aet_hot_wet <- global(mask(aet_2022, highest_aet_hot_wet), fun = "mean", na.rm = TRUE)
mean_cwd_hot_wet <- global(mask(cwd_2022, lowest_cwd_hot_wet), fun = "mean", na.rm = TRUE)

### Warm Dry

highest_aet_warm_dry <- selectHighest(aet_warm_dry, n = n_sites)
lowest_cwd_warm_dry <- selectHighest(cwd_warm_dry, n = n_sites, low = TRUE)

best_sites_warm_dry <- highest_aet_warm_dry & lowest_cwd_warm_dry

n_best_sites_warm_dry <- zonal(best_sites_warm_dry, best_sites_warm_dry, length)[[2]]

## (n_best_sites / ncell(elev)) * 100

mean_aet_warm_dry <- global(mask(aet_2022, highest_aet_warm_dry), fun = "mean", na.rm = TRUE)
mean_cwd_warm_dry <- global(mask(cwd_2022, lowest_cwd_warm_dry), fun = "mean", na.rm = TRUE)


### Hot Dry

highest_aet_hot_dry <- selectHighest(aet_hot_dry, n = n_sites)
lowest_cwd_hot_dry <- selectHighest(cwd_hot_dry, n = n_sites, low = TRUE)

best_sites_hot_dry <- highest_aet_hot_dry & lowest_cwd_hot_dry

n_best_sites_hot_dry <- zonal(best_sites_hot_dry, best_sites_hot_dry, length)[[2]]

## (n_best_sites / ncell(elev)) * 100

mean_aet_hot_dry <- global(mask(aet_2022, highest_aet_hot_dry), fun = "mean", na.rm = TRUE)
mean_cwd_hot_dry <- global(mask(cwd_2022, lowest_cwd_hot_dry), fun = "mean", na.rm = TRUE)


writeRaster(best_sites_historical, "raster/selected_sites_historical.tif", overwrite = TRUE)
writeRaster(best_sites_hot_dry, "raster/selected_sites_hot_dry.tif", overwrite = TRUE)
writeRaster(best_sites_hot_wet, "raster/selected_sites_hot_wet.tif", overwrite = TRUE)


best_sites_historical_plot <- ggplot() +
  geom_spatraster(data = shade, show.legend = FALSE) +
  ##    geom_spatraster(data = subset(tmmn$daily_minimum_temperature, 1)) +
  scale_fill_gradientn(colors = pal_greys, na.value = NA) +
  ggnewscale::new_scale_fill() +
  geom_spatraster(data = best_sites_historical) +
  scale_fill_discrete(na.value = "transparent") +
  geom_sf(data = fs_roads, color = "orange", lwd = 1.5, alpha = 0.5) +
  geom_sf_text(data = fs_roads, aes(label = ID), alpha = 0.9) +
  geom_sf(data = units, fill = NA, lwd = 1.5) +
  geom_sf_label(data = units, aes(label = Unit), alpha = 0.4) +
  labs(title = glue::glue("Historical (2002-2022), n = {n_best_sites_historical}, {round(n_best_sites_historical * 100 / ncell(elev), 2)}%")) +
  annotation_north_arrow(aes(location = "tr")) +
  annotation_scale(aes(location = "br"))

best_sites_warm_wet_plot <- ggplot() +
  geom_spatraster(data = shade, show.legend = FALSE) +
  ##    geom_spatraster(data = subset(tmmn$daily_minimum_temperature, 1)) +
  scale_fill_gradientn(colors = pal_greys, na.value = NA) +
  ggnewscale::new_scale_fill() +
  geom_spatraster(data = best_sites_warm_wet) +
  scale_fill_discrete(na.value = "transparent") +
  geom_sf(data = fs_roads, color = "orange", lwd = 1.5, alpha = 0.5) +
  geom_sf_text(data = fs_roads, aes(label = ID), alpha = 0.9) +
  geom_sf(data = units, fill = NA, lwd = 1.5) +
  geom_sf_label(data = units, aes(label = Unit), alpha = 0.4) +
  labs(title = glue::glue("Warm Wet (2075-2099), n = {n_best_sites_warm_wet}, {round(n_best_sites_warm_wet * 100 / ncell(elev), 2)}%")) +
  annotation_north_arrow(aes(location = "tr")) +
  annotation_scale(aes(location = "br"))

best_sites_hot_wet_plot <- ggplot() +
  geom_spatraster(data = shade, show.legend = FALSE) +
  ##    geom_spatraster(data = subset(tmmn$daily_minimum_temperature, 1)) +
  scale_fill_gradientn(colors = pal_greys, na.value = NA) +
  ggnewscale::new_scale_fill() +
  geom_spatraster(data = best_sites_hot_wet) +
  scale_fill_discrete(na.value = "transparent") +
  geom_sf(data = fs_roads, color = "orange", lwd = 1.5, alpha = 0.5) +
  geom_sf_text(data = fs_roads, aes(label = ID), alpha = 0.9) +
  geom_sf(data = units, fill = NA, lwd = 1.5) +
  geom_sf_label(data = units, aes(label = Unit), alpha = 0.4) +
  labs(title = glue::glue("Hot Wet (2075-2099), n = {n_best_sites_hot_wet}, {round(n_best_sites_hot_wet * 100 / ncell(elev), 2)}%")) +
  annotation_north_arrow(aes(location = "tr")) +
  annotation_scale(aes(location = "br"))

best_sites_warm_dry_plot <- ggplot() +
  geom_spatraster(data = shade, show.legend = FALSE) +
  ##    geom_spatraster(data = subset(tmmn$daily_minimum_temperature, 1)) +
  scale_fill_gradientn(colors = pal_greys, na.value = NA) +
  ggnewscale::new_scale_fill() +
  geom_spatraster(data = best_sites_warm_dry) +
  scale_fill_discrete(na.value = "transparent") +
  geom_sf(data = fs_roads, color = "orange", lwd = 1.5, alpha = 0.5) +
  geom_sf_text(data = fs_roads, aes(label = ID), alpha = 0.9) +
  geom_sf(data = units, fill = NA, lwd = 1.5) +
  geom_sf_label(data = units, aes(label = Unit), alpha = 0.4) +
  labs(title = glue::glue("Warm_Dry (2075-2099), n = {n_best_sites_warm_dry}, {round(n_best_sites_warm_dry * 100 / ncell(elev), 2)}%")) +
  annotation_north_arrow(aes(location = "tr")) +
  annotation_scale(aes(location = "br"))

best_sites_hot_dry_plot <- ggplot() +
    geom_spatraster(data = shade, show.legend = FALSE) +
    ##    geom_spatraster(data = subset(tmmn$daily_minimum_temperature, 1)) +
    scale_fill_gradientn(colors = pal_greys, na.value = NA) +
    ggnewscale::new_scale_fill() +
    geom_spatraster(data = best_sites_hot_dry) +
    scale_fill_discrete(na.value = "transparent") +
    geom_sf(data = fs_roads, color = "orange", lwd = 1.5, alpha = 0.5) +
    geom_sf_text(data = fs_roads, aes(label = ID), alpha = 0.9) +
    geom_sf(data = units, fill = NA, lwd = 1.5) +
    geom_sf_label(data = units, aes(label = Unit), alpha = 0.4) +
    labs(title = glue::glue("Hot_Dry (2075-2099), n = {n_best_sites_hot_dry}, {round(n_best_sites_hot_dry * 100 / ncell(elev), 2)}%")) +
    annotation_north_arrow(aes(location = "tr")) +
    annotation_scale(aes(location = "br"))

```

``` {r best-plantings-historical, eval = F, include = F, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 10, fig.width = 10, fig.cap = "Locations that maximize AET and CWD for historical (2002-2022). n = number of pixels selected by algorithm, percentage shown is percentage of landscape identified by algorithm."}

best_sites_historical_plot

```

``` {r best-plantings-projections, include = F, eval = F, warning = FALSE, message = FALSE, echo = FALSE, fig.height = 16, fig.width = 16, fig.cap = "Locations that maximize AET and CWD for RCP4.5/RCP8.5 end-of-century projection ensembles. n = number of pixels selected by algorithm, percentage shown is percentage of landscape identified by algorithm."}
ggarrange(best_sites_warm_wet_plot, best_sites_hot_wet_plot,
          best_sites_warm_dry_plot, best_sites_hot_dry_plot,
          common.legend = TRUE,
          nrow = 2,
          ncol = 2
          ) %>%
    annotate_figure(., top = text_grob("Selected Planting Sites", size = 14))

```

This algorithm identifies areas of the landscape that are ideal based on the criteria of maximizing AET and minimizing CWD, relative to other pixels in the landscape for the same time period.  Much of the area identified by this algorithm in the based on conditions over the past twenty years is likely to remain the areas that maximize AET and minimize CWD under projected climatic conditions at the end of the century (Figures \@ref(fig:best-plantings-historical),  \@ref(fig:best-plantings-projections)).  However, widespread warming under climate projections will also affect these areas, and they will likely experience increases in both annual AET and Deficit by the end of the century as will the rest of the landscape.  

A large number of pixels is needed because maximizing AET and minimizing CWD is likely to select for riparian and other high-productivity areas that may not be suitable or competitive growing areas for WBP.  Selecting a large number of points will select for these areas that are likely to not be suitable for WBP as well as more mesic locations that are potentially suitable. 

Average conditions of "ideal pixels" (red pixels in Figures \@ref(fig:best-plantings-historical),  \@ref(fig:best-plantings-projections)) in historic, RCP4.5 and RCP8.5 scenarios:

| Scenario                                 | Mean AET                | Mean CWD                |
|------------------------------------------|-------------------------|-------------------------|
| Historic                                 | `r mean_aet_historical` | `r mean_cwd_historical` |
| Warm Wet 2075-2099 (MRI-CGCM3 RCP8.5)    | `r mean_aet_warm_wet`   | `r mean_cwd_warm_wet`   |
| Hot Wet 2075-2099 (CanESM2 RCP8.5)       | `r mean_aet_hot_wet`    | `r mean_cwd_hot_wet`    |
| Warm Dry 2075-2099 (MRI-CGCM3 RCP4.5)    | `r mean_aet_warm_dry`   | `r mean_cwd_warm_dry`   |
| Hot Dry 2075-2099 (HadGEM2-CC365 RCP8.5) | `r mean_aet_hot_dry`    | `r mean_cwd_hot_dry`    |



## Climate Summaries

Table: (\#tab:change-from-baseline-table) Change from AET and CWD baseline (2002-2022) in RCP4.5 and RCP8.5 ensembles
  
| Scenario                                 | AET Mean change from 2002-2022 Baseline          | CWD Mean change from 2002-2022 Baseline          |
|------------------------------------------|--------------------------------------------------|--------------------------------------------------|
| Warm Wet 2075-2099 (MRI-CGCM3 RCP8.5)    | `r global(mean(aet_warm_wet_change), na.rm = TRUE)`            | `r global(mean(cwd_warm_wet_change), na.rm = TRUE)`            |
| Hot Wet 2075-2099 (CanESM2 RCP8.5)       | `r global(mean(aet_hot_wet_change), na.rm = TRUE)`            | `r global(mean(cwd_hot_wet_change), na.rm = TRUE)`             |
| Warm Dry 2075-2099 (MRI-CGCM3 RCP4.5)    | `r global(mean(aet_warm_dry_change), na.rm = TRUE)`            | `r global(mean(cwd_warm_dry_change), na.rm = TRUE)`            |
| Hot Dry 2075-2099 (HadGEM2-CC365 RCP8.5) | `r global(mean(aet_hot_dry_change), na.rm = TRUE)`            | `r global(mean(cwd_hot_dry_change), na.rm = TRUE)`             |

Table: (\#tab:climate-summary) Summary of climate conditions for historic data and projections.

| Year                                     | Mean AET                       | Mean CWD                       |
|------------------------------------------|--------------------------------|--------------------------------|
| 1988 (Drought Year)                      | `r global(mean(aet_1988), na.rm = TRUE)`     | `r global(mean(cwd_1988), na.rm = TRUE)`     |
| 2013 (Fire Year)                      | `r global(mean(aet_2013), na.rm = TRUE)`     | `r global(mean(cwd_2013), na.rm = TRUE)`     |
| 1982-2002                                | `r global(mean(aet_2002), na.rm = TRUE)`     | `r global(mean(cwd_2002), na.rm = TRUE)`     |
| 2002-2022                                | `r global(mean(aet_2022), na.rm = TRUE)`     | `r global(mean(cwd_2022), na.rm = TRUE)`     |
| Warm Wet 2075-2099 (MRI-CGCM3 RCP8.5)    | `r global(mean(aet_warm_wet), na.rm = TRUE)` | `r global(mean(cwd_warm_wet), na.rm = TRUE)` |
| Hot Wet 2075-2099 (CanESM2 RCP8.5)       | `r global(mean(aet_hot_wet), na.rm = TRUE)`  | `r global(mean(cwd_hot_wet), na.rm = TRUE)`  |
| Warm Dry 2075-2099 (MRI-CGCM3 RCP4.5)    | `r global(mean(aet_warm_dry), na.rm = TRUE)` | `r global(mean(cwd_warm_dry), na.rm = TRUE)` |
| Hot Dry 2075-2099 (HadGEM2-CC365 RCP8.5) | `r global(mean(aet_hot_dry), na.rm = TRUE)`  | `r global(mean(cwd_hot_dry), na.rm = TRUE)`  |

# Limitations
Snow accumulation and melt were modeled at the scale of the entire
site, and was not downscaled based on topography.  In reality, snow
accumulation and melt is affected by factors such as slope, aspect, as
well as movement through snow drift.  Future analyses could improve
modeling of fine-scale water balance patterns by incorporating
modeling of snow melt and accumulation dynamics at fine spatial scales

Temperature and precipitation data were applied uniformly across all
pixels in the site.  However, temperature is affected by changes in
elevation and this can impact the accuracy of modeled water balance
metrics.  Future analyses could incorporate temperature corrections
based on lapse rate to improve accuracy of modeled metrics that are
temperature-sensitive [@tercekCorrectlyApplyingLapse2021].

Soil survey data provided by the SSURGO database is intended to be
used at scales of 1:24,000.  There is likely considerable variation in
the soil characteristics observed at the fine spatial scales analyzed
here that is not captured in these surveys.  Planting decisions made
using this data should consider observed soil characteristics in the
field in conjunction with data presented here.

This model is relative in nature, and absolute values from this model
cannot be compared with values generated by water balance models using
different parameters (spatial scale, soil WHC, input DEM).  However,
relative patterns are comparable between models at different scales.
For example, the relative dryness of a position on the landscape
compared to other positions should be comparable between water balance
models at different scales, for example the 1 m water balance model 
presented here and the 1 km NPS gridded water balance product [@tercekHistoricalChangesPlant2021].  


# Discussion

The Burroughs Creek site presents opportunities to take advantage of
microsites created by the effects of local topographic and soil
features.  Given that the projected climate 

Extreme CWD is likely to threaten tree survival on the site in the
future.  In 1988, a year in which severe drought led to widespread
wildfire through the greater region, the Burroughs Creek site (defined
by the bounding box surrounding the two planting units) experienced an
average annual CWD of 269.93 mm (Table \@ref(tab:climate-summary)),
with some locations experiencing higher or lower deficit depending on
the effects of topoedaphic features (Figure
\@ref(fig:drought-year-plot)). All GCMs examined here show average annual deficits approaching or exceeding the deficit during the 1988 drought \@ref(tab:climate-summary).  It is likely that
plants on the Burroughs Creek site will experience severe drought
stress in the future, leading to increase mortality as well as
likelihood of wildfire ignition [@thomaWaterBalanceIndicator2020].  To
maximize success of tree plantings on this site, planters can utilize
topoedaphic features to plant in locations that buffer from extreme
CWD, which will reduce tree stress and promote survival of planted
seedlings.  

Optimizing for AET in planting locations would increase seedling
growth rates and establishment
[@laufenbergBiophysicalGradientsPerformance2020]. In this landscape,
differences in AET appear to be mainly driven by differences in soil
WHC, which sustains evapotranspiration for longer in areas with higher
plant available water storage (Figures
\@ref(fig:historical-aet-plot),\@ref(fig:projections-aet), \@ref(fig:projections-cwd)). North
facing slopes which reduce observed CWD also have the effect of
decreasing AET due to reduced incoming solar radiation.  Given the
extreme levels of CWD projected on this location, drought mortality in
planted trees is likely to be high and planters may wish to select
planting locations that minimize drought stress at the expense of
slower seedling growth rates.

Unit RL1 is mostly homogeneous in topoedaphic features, being
primarily an open face that is SE-facing and characterized by low to
moderate soil WHC for the area (Tables
\@ref(tab:planting-unit-statistics), \@ref(tab:soils-zonals)).
Planting microclimates in unit RL1 are likely to be similar throughout
the unit, however, some landscape features are present which could
create favorable planting microclimates such as the small hill in the
south section of the unit where the northern side of the hill appears
to buffer the location from extreme values of deficit (Figures \@ref(fig:best-plantings-historical),  \@ref(fig:best-plantings-projections)).

Unit WB2 is more diverse in soils and topography, with deeper soils
and more rugged terrain than unit RL1 (Tables
\@ref(tab:planting-unit-statistics), \@ref(tab:soils-zonals)).  This
creates a broader range of microclimates, with a broader range of
observed AET and CWD throughout the unit when compared to RL1 (Figures
\@ref(fig:historical-aet-plot), \@ref(fig:historical-cwd-plot),
\@ref(fig:drought-year-plot), \@ref(fig:projections-aet),
\@ref(fig:projections-cwd).  Undulations in terrain in this unit
create a variety of north to east facing slopes and shelves that can
create suitable microclimates for planting (Figures \@ref(fig:best-plantings-historical),  \@ref(fig:best-plantings-projections)), while more exposed locations with southern
aspects show some of the most extreme CWD values observed on the site
(Figure \@ref(fig:projections-cwd)).  These same areas that create
favorable conditions measured by annual AET and CWD also appear to
limit interannual variability in CWD projections (Figure
\@ref(fig:projections-sd-cwd)), promoting tree longevity, and minimize
change in CWD and AET from baseline conditions (Figures
\@ref(fig:projections-change-cwd), \@ref(fig:projections-change-aet)).  

Unit WB2 has large areas of more hydric soils with higher WHC present
at the bottom of drainages in the unit (Figure \@ref(fig:soils)). The
relatively higher soil WHC present in these soils has the effect of
increasing AET and decreasing CWD where they are present.  However,
these soils currently support riparian plant communities (soils
report), and are not likely to be locations that whitebark seedlings
could be competitive. Due to limitations in the available soil data,
these soils are treated as relatively discrete areas, however, in
reality there is likely to be considerable gradation between and
heterogeneity within soil types. Around the edges of these more hydric
soils, it is likely that locations can be found that have more plant
available water than would be expected based on this modeled data. *In
situ*, trees can also access features like subsurface flows and water
in bedrock [@mccormickWidespreadWoodyPlant2021], which is not
accounted for in this model.

A large portion of the area that was identified as optimal for
planting (Figures \@ref(fig:best-plantings-historical),  \@ref(fig:best-plantings-projections)) is present outside of the
two planting units. Satellite imagery shows existing tree cover in
most of the area outside of the planting unit boundaries.  However, if
planting materials are available and openings exist in some of the
areas, planters may wish to plant outside of the boundaries where
suitable planting microclimates exist.

<!-- # TODO -->

<!-- - 3d Maps -->
<!-- - Other ways to identify ideal planting locations.  What criteria are Nancy/planting crews looking for? -->

```{r, rayshader, include = F, eval = F
library(rayshader)

cwd_2002_2022_rgb <- rast("raster/cwd_2002-2022.tif")

elev_crop <- crop(elev, cwd_2002_2022_rgb)

# breakout RGB layers
names(cwd_2002_2022_rgb) <- c("r", "g", "b", "x")
topo_r <- rayshader::raster_to_matrix(cwd_2002_2022_rgb$r)
topo_g <- rayshader::raster_to_matrix(cwd_2002_2022_rgb$g)
topo_b <- rayshader::raster_to_matrix(cwd_2002_2022_rgb$b)
topo_rgb_array <- array(0, dim = c(nrow(topo_r), ncol(topo_r), 3))
topo_rgb_array[,,1] <- topo_r/255
topo_rgb_array[,,2] <- topo_g/255
topo_rgb_array[,,3] <- topo_b/255
topo_rgb_array <- aperm(topo_rgb_array, c(2,1,3))

dims <- dim(topo_rgb_array)
width = dims[2]
height = dims[1]

zscale = 1

# rayshade
elev_mat <- raster_to_matrix(elev_crop)
elev_mat = resize_matrix(elev_mat, scale=1)
ray_shadow <- ray_shade(elev_mat, sunaltitude = 30, zscale = zscale, multicore = TRUE)
ambient_shadow <- ambient_shade(elev_mat, zscale = zscale)

elev_mat %>%
  sphere_shade() %>%
  add_overlay(topo_rgb_array, rescale_original = TRUE) %>%
  add_water(detect_water(elev_mat, min_area = 400), color = "paleturquoise") %>%
  #add_shadow(ray_shadow, max_darken = 0.3) %>%
  #add_shadow(ambient_shadow, 0.3) %>%
  #add_overlay(generate_compass_overlay(heightmap = elev_mat)) %>%
  #add_overlay(generate_scalebar_overlay(extent = ext(elev), length = 100, heightmap = elev_mat)) %>%
  ##add_overlay(generate_polygon_overlay(poly, extent = ext(elev), heightmap = elev_mat)) %>%
  plot_3d(heightmap = elev_mat,
          zscale = zscale,
          solid = TRUE,
          water = TRUE,
          waterdepth = 1,
          fov = 0,
          theta = 45,
          zoom = 0.75,
          phi = 45,
          windowsize = c(1280, 1280),
          background = "white")

render_movie("raster/burroughs.mp4", theta = 45, phi = 45, zoom = 0.75, fov = 0,
             frames = 1080,
             fps = 60,
             height = 1280,
             width = 1280)

render_camera(theta=0,phi=90,fov=60,zoom=0.6)

render_snapshot("snapshot.png",
                software_render = TRUE,
                width = width,
                height = height)

### Save as object to render in browser
save_obj("raster/cwd_2002_2022.obj", save_texture = TRUE)
```

# Acknowledgments
Computational efforts were performed on the [Tempest High Performance
Computing System](https://www.montana.edu/uit/rci/tempest/), operated and supported by University Information
Technology Research Cyberinfrastructure at Montana State University.

# References

<!-- Local Variables: -->
<!-- jinx-local-words: "ncell" -->
<!-- End: -->


```{js, echo=FALSE}
var elements = document.getElementsByTagName('img');
for(var i = 0, len = elements.length; i < len; i++) {
    elements[i].onclick = function () {
        window.open(this.src, '_blank');
    }
}
```
